+++ packages/doctor/src/getRepeatabilityReports.ts
@@ -1,7 +1,7 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
-// import { IClone } from '@jscpd/core';
-// import { jscpd } from 'jscpd';
+import { IClone } from '@jscpd/core';
+import { jscpd } from 'jscpd';
 import Scorer from './Scorer';
 import { IRepeatabilityReports } from './types/Scanner';
 
@@ -12,37 +12,37 @@ const tempDir = path.join(__dirname, 'tmp/');
 export default async function getRepeatabilityReports(
   directory: string,
   supportExts: string[],
-  ignoreDirs: string[]
+  ignore: string[]
 ): Promise<IRepeatabilityReports> {
-  // let clones: IClone[] = [];
-  // let repetitionPercentage = 0;
+  let clones: IClone[] = [];
+  let repetitionPercentage = 0;
 
-  // try {
-  //   clones = await jscpd([
-  //     '--formats-exts',
-  //     supportExts.join(','),
-  //     directory,
-  //     '--ignore',
-  //     `"${ignoreDirs.map((ignoreDir) => `${path.join(directory, '/')}**/${ignoreDir}/**`).join(',')}"`,
-  //     '--reporters',
-  //     'json',
-  //     '--output',
-  //     tempDir,
-  //     '--silent',
-  //   ]);
+  try {
+    clones = await jscpd([
+      '--formats-exts',
+      supportExts.join(','),
+      directory,
+      '--ignore',
+      `"${ignore.map((ignoreDir) => `${path.join(directory, '/')}**/${ignoreDir}/**`).join(',')}"`,
+      '--reporters',
+      'json',
+      '--output',
+      tempDir,
+      '--silent',
+    ]);
 
-  //   const repeatabilityResultFile = path.join(tempDir, 'jscpd-report.json');
-  //   if (fs.existsSync(repeatabilityResultFile)) {
-  //     const repeatabilityResult = fs.readJSONSync(repeatabilityResultFile);
-  //     repetitionPercentage = repeatabilityResult.statistics.total.percentage;
-  //   }
-  // } catch (e) {
-  //   // ignore
-  //   console.log(e);
-  // }
+    const repeatabilityResultFile = path.join(tempDir, 'jscpd-report.json');
+    if (fs.existsSync(repeatabilityResultFile)) {
+      const repeatabilityResult = fs.readJSONSync(repeatabilityResultFile);
+      repetitionPercentage = repeatabilityResult.statistics.total.percentage;
+    }
+  } catch (e) {
+    // ignore
+    console.log(e);
+  }
 
   return {
-    score: 0,
-    clones: [],
+    score: new Scorer().minus(repetitionPercentage),
+    clones,
   };
 }