+++ extensions/iceworks-app/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 
-# 0.4.8
+## 0.4.10
+
+- docs: update text for find component and generate page
+
+## 0.4.9
+
+- feat: add create component to quick entry
+
+## 0.4.8
+
 - fix: build error
 
 # 0.4.7