+++ extensions/iceworks-material-helper/CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 
-# 0.3.3
+## 0.3.5
+
+- docs: en-US demo
+- refactor: remove page logic
+
+## 0.3.4
+
+- fix: generate block code error
+
+## 0.3.3
 
 - fix: build error