+++ extensions/iceworks-project-creator/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 0.3.10
+
+- feat: add Rax scaffolds to external
+
+## 0.3.9
+
+- fix: build error
+
 ## 0.3.8
 
 - feat: update pegasus project git clone action.