+++ extensions/iceworks-ui-builder/CHANGELOG.md
@@ -1,6 +1,25 @@
 # Change Log
 
-# 0.1.9
+## 0.1.13
+
+- docs: update text for find component and generate page
+
+## 0.1.12
+
+- fix: fix form style.
+
+## 0.1.11
+
+- feat: show open file action after create component
+- refactor: router logic
+- fix: fail to generate text style in component-generator 
+- fix: fail to preview component attributes in component-generator canvas panel
+
+## 0.1.10
+
+- fix: generate block code error
+
+## 0.1.9
 
 - fix: build error